Refactor (gguf): moving proxy and caching utils into separate files
